+++ inertia/components/layout/intervention/EditWorkdoneForm.tsx
@@ -194,7 +194,6 @@ export default function EditWorkDoneForm({
 				value={data.work_done}
 				onChange={(e) => setData("work_done", e.target.value)}
 				error={errors.work_done}
-				placeholder="Travaux effectués"
 				required
 			/>
 
@@ -203,7 +202,6 @@ export default function EditWorkDoneForm({
 				value={data.used_materials as string}
 				onChange={(e) => setData("used_materials", e.target.value)}
 				error={errors.used_materials}
-				placeholder="Matériel utilisé"
 			/>
 
 			<Input